+++ test/plug_rest/router_test.exs
@@ -253,11 +253,6 @@ defmodule PlugRest.RouterTest do
     use PlugRest.Resource
 
     def to_html(conn, state) do
-      ## Accesing path params through conn.params is deprecated
-      _user_id = conn.params["user_id"]
-      _comment_id = conn.params["comment_id"]
-
-      ## Use read_path_params(conn) instead
       params = read_path_params(conn)
       user_id = params["user_id"]
       comment_id = params["comment_id"]